chore(test): normalize test file layout
